Put init and de-init Shizuku to the domain module
